pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.163.0",
    "google-auth>=2.38.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "streamlit>=1.43.1",
]
